Add files via upload
better formatted scrapes. Works with weapons page so far, testing/adjustment for other pages to come. Removes extraneous data at the end of the output that used to be present. There is now a capability to add custom filters to only get tables from specific header strings. Some tables that are formatted differently (e.g. "Setting Specific Weapons" on the weapons page) still aren't working, they may just need to be addressed individually. 
Add all observed bugs/notes/to-do's to log.
Json return not currently working with this version, I'm still building the functionality in so that it works systematically. Will create and upload an example .json of the new formatting.
Currently this will also just print the output from the weapons page scrape, so it can be seen that way.